[BREAKING] consolidate on a simplified `reduce`
This removes the many variants of `reduce` in favor of just one which
takes two closures. This is equally general.  I kept the `ReduceOp`
trait in the implementation because I don't think there is an easy way
to handle methods like `sum` and `min` and `max` otherwise, since
we lack a generic way to produce identity.